+++ pkg/mads/generator/assignments_test.go
@@ -52,8 +52,10 @@ var _ = Describe("MonitoringAssignmentsGenerator", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{{
-									Interface: "192.168.0.1:80:8080",
+									Port:        80,
+									ServicePort: 8080,
 									Tags: map[string]string{
 										"service": "backend",
 									},
@@ -81,8 +83,10 @@ var _ = Describe("MonitoringAssignmentsGenerator", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{{
-									Interface: "192.168.0.1:80:8080",
+									Port:        80,
+									ServicePort: 8080,
 									Tags: map[string]string{
 										"service": "backend",
 									},
@@ -186,9 +190,11 @@ var _ = Describe("MonitoringAssignmentsGenerator", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
 									{
-										Interface: "192.168.0.1:80:8080",
+										Port:        80,
+										ServicePort: 8080,
 										Tags: map[string]string{
 											"service": "backend",
 											"env":     "prod",
@@ -196,7 +202,9 @@ var _ = Describe("MonitoringAssignmentsGenerator", func() {
 										},
 									},
 									{
-										Interface: "192.168.0.2:443:8443",
+										Address:     "192.168.0.2",
+										Port:        443,
+										ServicePort: 8443,
 										Tags: map[string]string{
 											"service": "backend-https",
 											"env":     "prod",
@@ -262,8 +270,10 @@ var _ = Describe("MonitoringAssignmentsGenerator", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{{
-									Interface: "192.168.0.1:80:8080",
+									Port:        80,
+									ServicePort: 8080,
 									Tags: map[string]string{
 										"service":  "backend",
 										"version":  "v1",
@@ -327,8 +337,10 @@ var _ = Describe("MonitoringAssignmentsGenerator", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{{
-									Interface: "192.168.0.1:80:8080",
+									Port:        80,
+									ServicePort: 8080,
 									Tags: map[string]string{
 										"service":         "backend",
 										"app:description": "?!,.:;",
@@ -407,8 +419,10 @@ var _ = Describe("MonitoringAssignmentsGenerator", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{{
-									Interface: "192.168.0.1:80:8080",
+									Port:        80,
+									ServicePort: 8080,
 									Tags: map[string]string{
 										"service": "backend",
 										"env":     "prod",
@@ -424,8 +438,10 @@ var _ = Describe("MonitoringAssignmentsGenerator", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.2",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{{
-									Interface: "192.168.0.2:443:8443",
+									Port:        443,
+									ServicePort: 8443,
 									Tags: map[string]string{
 										"service": "web",
 										"env":     "intg",