+++ pkg/xds/generator/proxy_template_raw_source_test.go
@@ -50,8 +50,12 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
-									{Interface: "192.168.0.1:80:8080"},
+									{
+										Port:        80,
+										ServicePort: 8080,
+									},
 								},
 							},
 						},
@@ -74,8 +78,12 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
-									{Interface: "192.168.0.1:80:8080"},
+									{
+										Port:        80,
+										ServicePort: 8080,
+									},
 								},
 							},
 						},
@@ -97,8 +105,12 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
-									{Interface: "192.168.0.1:80:8080"},
+									{
+										Port:        80,
+										ServicePort: 8080,
+									},
 								},
 							},
 						},
@@ -122,8 +134,12 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
-									{Interface: "192.168.0.1:80:8080"},
+									{
+										Port:        80,
+										ServicePort: 8080,
+									},
 								},
 							},
 						},
@@ -158,8 +174,12 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
-									{Interface: "192.168.0.1:80:8080"},
+									{
+										Port:        80,
+										ServicePort: 8080,
+									},
 								},
 							},
 						},
@@ -229,8 +249,12 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
-									{Interface: "192.168.0.1:80:8080"},
+									{
+										Port:        80,
+										ServicePort: 8080,
+									},
 								},
 							},
 						},
@@ -248,8 +272,12 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
-									{Interface: "192.168.0.1:80:8080"},
+									{
+										Port:        80,
+										ServicePort: 8080,
+									},
 								},
 							},
 						},
@@ -305,8 +333,12 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
-									{Interface: "192.168.0.1:80:8080"},
+									{
+										Port:        80,
+										ServicePort: 8080,
+									},
 								},
 							},
 						},
@@ -360,8 +392,12 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 						},
 						Spec: mesh_proto.Dataplane{
 							Networking: &mesh_proto.Dataplane_Networking{
+								Address: "192.168.0.1",
 								Inbound: []*mesh_proto.Dataplane_Networking_Inbound{
-									{Interface: "192.168.0.1:80:8080"},
+									{
+										Port:        80,
+										ServicePort: 8080,
+									},
 								},
 							},
 						},